+++ ElvUI/Game/Shared/Modules/Auras/Containers.lua
@@ -33,6 +33,16 @@ E.AuraHighlight = {
  -- customDispelColorCurve is added from UpdateAuraCurves
 }
 
+E.AuraGroupHeaders = {
+	party = true,
+	raid1 = true,
+	raid2 = true,
+	raid3 = true,
+	raidpet = true,
+	assist = true,
+	tank = true,
+}
+
 E.AuraEventUnits = {
 	PLAYER_TARGET_CHANGED = 'target',
 	PLAYER_FOCUS_CHANGED = 'focus'
@@ -85,7 +95,7 @@ if SORTDIRECTION then
 	E.AuraContainerSortDirection['-'] = SORTDIRECTION.Reverse
 end
 
-function E:Auras_OnEvent(event, arg1, arg2)
+function E:Auras_OnEvent(event, arg1)
 	local container = self.owner
 	if event == 'PLAYER_FOCUS_CHANGED' or event == 'PLAYER_TARGET_CHANGED' then
 		local eventUnit = E.AuraEventUnits[event]
@@ -94,13 +104,19 @@ function E:Auras_OnEvent(event, arg1, arg2)
 				UF:AuraBars_UpdateFilter(container, eventUnit)
 				E:Auras_SetContainer(container)
 			else -- for target frame
+				E:Auras_AssistUnit(container, eventUnit)
+
 				container:UpdateAllAuras()
 			end
 		end
-	elseif arg1 and (arg1 == container.unit) then -- about to do something
-		if event == 'UNIT_FACTION' or event == 'UNIT_TARGETABLE_CHANGED' then
+	elseif event == 'GROUP_ROSTER_UPDATE' then
+		if container.unit and E.AuraGroupHeaders[container.unitframeType] then
+			E:Auras_AssistUnit(container, container.unit)
+
 			container:UpdateAllAuras()
 		end
+	elseif arg1 and (arg1 == container.unit) then
+		E:Auras_AssistUnit(container, arg1)
 	end
 end
 
@@ -132,6 +148,51 @@ function E:Auras_UpdateHighlight(container, button)
 	end
 end
 
+function E:Auras_CreateText(button)
+	local textFrame = CreateFrame('Frame', nil, button)
+	textFrame:SetAllPoints()
+
+	local countText = textFrame:CreateFontString(nil, 'OVERLAY')
+	textFrame.count = countText
+
+	local timeText = textFrame:CreateFontString(nil, 'OVERLAY')
+	timeText:FontTemplate(nil, 14)
+	timeText:Point('CENTER')
+	textFrame.time = timeText
+
+	local nameText = textFrame:CreateFontString(nil, 'OVERLAY')
+	nameText:Point('LEFT', button, 2, 0)
+	textFrame.nameText = nameText
+
+	return textFrame
+end
+
+function E:Auras_UpdateText(container, button)
+	local text = button.textFrame
+	if not text then return end
+
+	local count = text.count
+	if count then
+		local data = button.data -- indicators
+		local point = (data and data.countAnchor) or container.countPosition or (data and 'TOPRIGHT' or 'CENTER')
+		local x = (data and data.countX) or container.countXOffset or (data and 1 or 0)
+		local y = (data and data.countY) or container.countYOffset or (data and 1 or 0)
+
+		count:ClearAllPoints()
+		count:Point(point, x, y)
+		count:SetJustifyH(strfind(point, 'RIGHT') and 'RIGHT' or 'LEFT')
+		count:FontTemplate(container.countFont, container.countFontSize or (data and 12 or nil), container.countFontOutline or (data and 'OUTLINE' or nil))
+
+		button:SetApplicationCount(count)
+	end
+
+	local name = container.isAuraBar and text.nameText
+	if name then
+		name:FontTemplate(container.textFont, container.textFontSize, container.textFontOutline)
+		button:SetSpellName(name)
+	end
+end
+
 function E:Auras_CreateIndicator(button)
 	button:CreateBackdrop('Transparent', nil, true) -- these are forbidden, ignore updates
 
@@ -145,6 +206,8 @@ function E:Auras_CreateIndicator(button)
 	local cooldown = CreateFrame('Cooldown', nil, button, 'CooldownFrameTemplate')
 	cooldown:SetAllPoints(texture)
 	button.cooldown = cooldown
+
+	button.textFrame = E:Auras_CreateText(button)
 end
 
 function E:Auras_UpdateIndicator(container, button)
@@ -234,6 +297,10 @@ function E:Auras_UpdateIndicator(container, button)
 			end
 		end
 	end
+
+	if button.textFrame then
+		E:Auras_UpdateText(container, button)
+	end
 end
 
 function E:Auras_CreateButton(button)
@@ -277,23 +344,7 @@ function E:Auras_CreateButton(button)
 	cooldown:SetAllPoints(texture)
 	button.cooldown = cooldown
 
-	local textFrame = CreateFrame('Frame', nil, button)
-	if textFrame then
-		textFrame:SetAllPoints()
-		button.textFrame = textFrame
-
-		local countText = textFrame:CreateFontString(nil, 'OVERLAY')
-		textFrame.count = countText
-
-		local timeText = textFrame:CreateFontString(nil, 'OVERLAY')
-		timeText:FontTemplate(nil, 14)
-		timeText:Point('CENTER')
-		textFrame.time = timeText
-
-		local nameText = textFrame:CreateFontString(nil, 'OVERLAY')
-		nameText:Point('LEFT', button, 2, 0)
-		textFrame.nameText = nameText
-	end
+	button.textFrame = E:Auras_CreateText(button)
 end
 
 function E:Auras_UpdateButton(container, button)
@@ -381,7 +432,7 @@ function E:Auras_UpdateButton(container, button)
 		if container.useStatusbar then
 			button:SetDurationBar(statusbar)
 
-			local color = container.customBackdropColor or backdropColor
+			local color = container.barColor or backdropColor
 			statusbar:SetStatusBarTexture(container.barTexture)
 			statusbar:SetStatusBarColor(color.r or 1, color.g or 1, color.b or 1, container.isTransparent and backdropFadeColor.a or 1)
 			statusbar:Show()
@@ -445,24 +496,8 @@ function E:Auras_UpdateButton(container, button)
 		end
 	end
 
-	local textFrame = button.textFrame
-	if textFrame then
-		local count = textFrame.count
-		if count then
-			local point = container.countPosition or 'CENTER'
-			count:ClearAllPoints()
-			count:Point(point, container.countXOffset or 0, container.countYOffset or 0)
-			count:SetJustifyH(strfind(point, 'RIGHT') and 'RIGHT' or 'LEFT')
-			count:FontTemplate(container.countFont, container.countFontSize, container.countFontOutline)
-
-			button:SetApplicationCount(count)
-		end
-
-		local name = container.isAuraBar and textFrame.nameText
-		if name then
-			name:FontTemplate(container.textFont, container.textFontSize, container.textFontOutline)
-			button:SetSpellName(name)
-		end
+	if button.textFrame then
+		E:Auras_UpdateText(container, button)
 	end
 
 	if container.MasqueGroup then
@@ -484,6 +519,14 @@ function E:Auras_UpdateIndicators(container)
 
 	for button in next, container.indicators do
 		E:Auras_UpdateIndicator(container, button)
+	end
+end
+
+function E:Auras_UpdateHighlights(container)
+	if E:IsRestrictedAuras() then return end
+
+	for button in next, container.indicators do
+		E:Auras_UpdateHighlight(container, button)
 	end
 end
 
@@ -551,39 +594,21 @@ function E:Auras_UpdateLayout(container)
 end
 
 do
-	local temp = {}
-	local spell = {}
-	function E:Auras_FilterIndicator(data)
-		temp.includeSpellIDs = spell
-
+	local temp, spell = {}, {}
+	function E:Auras_FilterSlot(container, data)
 		wipe(spell)
 
-		local dataID = data.id
-		if dataID then
-			spell[dataID] = true
-		end
-
-		return temp
-	end
-end
-
-do
-	local spell = {}
-	function E:Auras_FilterHighlight(container, data)
-		local temp = container.candidateTemp
-		wipe(temp) -- trash object for reuse
-
 		if data then
-			temp.includeSpellIDs = spell
-
-			wipe(spell)
-
 			local dataID = data.id
 			if dataID then
 				spell[dataID] = true
 			end
+
+			temp.includeSpellIDs = spell
+			temp.includeDispelTypes = nil
 		else
-			temp.includeDispelTypes = E:CopyTable({}, DispelTypes)
+			temp.includeSpellIDs = nil
+			temp.includeDispelTypes = E:CopyTable(spell, DispelTypes)
 		end
 
 		return temp
@@ -672,11 +697,8 @@ function E:Auras_AddEnchantments(container)
 end
 
 function E:Auras_UpdateSlot(container, key, filter, candidate, sortMethod, sortDirection)
-	if candidate then
-		container:SetAuraSlotCandidateFilters(key, candidate)
-	end
-
 	container:SetAuraSlotFilterString(key, filter)
+	container:SetAuraSlotCandidateFilters(key, candidate)
 	container:SetAuraSlotSortMethod(key, sortMethod, sortDirection)
 end
 
@@ -685,30 +707,52 @@ function E:Auras_AddSlot(container, key, filter, candidate, sortMethod, sortDire
 	container:AddAuraSlot(key, filter, slot)
 end
 
+function E:Auras_HighlightFilter(container, data)
+	if data and data.enable then
+		return container.filter .. (data.ownOnly and '|PLAYER' or '')
+	end
+
+	return ''
+end
+
 function E:Auras_SetHighlight(container)
 	local groupKey = container.key
 	if groupKey == 'bad' then
 		if container.known[groupKey] then return end
 
-		local candidate = E:Auras_FilterHighlight(container)
-		container.candidateFilters = candidate
+		local candidate = E:Auras_FilterSlot(container)
+		E:Auras_CleanCandidates(container, candidate)
 
 		local slot = E:Auras_SetupHighlight(container, candidate)
 		container:AddAuraSlot(groupKey, container.filter, slot)
 
 		container.known[groupKey] = 'meow'
 	else
-		for key, data in next, container.keys do
-			local candidate = E:Auras_FilterHighlight(container, data)
-			container.candidateFilters = candidate
+		for key, data in next, container.active do
+			if not container.keys[key] then -- only handle previous keys
+				local filter = E:Auras_HighlightFilter(container, data)
+				if filter ~= '' then -- it was something else so lets turn it off
+					container:SetAuraSlotFilterString(key, '')
+				end
+			end
 
-			local slotFilter = container.filter .. (data.ownOnly and '|PLAYER' or '')
+			container.active[key] = nil
+		end
+
+		for key, data in next, container.keys do
+			container.active[key] = data
+
+			local filter = E:Auras_HighlightFilter(container, data)
+			local candidate = E:Auras_FilterSlot(container, data)
+			E:Auras_CleanCandidates(container, candidate)
+
 			if container.known[key] then
-				container:SetAuraSlotFilterString(key, slotFilter)
+				container:SetAuraSlotFilterString(key, filter)
 				container:SetAuraSlotCandidateFilters(key, candidate)
 			else
 				local slot = E:Auras_SetupHighlight(container, candidate, key, data)
-				container:AddAuraSlot(key, slotFilter, slot)
+				container:AddAuraSlot(key, filter, slot)
+
 				container.known[key] = 'bark'
 			end
 		end
@@ -720,8 +764,8 @@ function E:Auras_SetIndicator(container)
 	local sortDirection = container.sortDirection or SORTDIRECTION.Normal
 
 	for key, data in next, container.keys do
-		local candidate = E:Auras_FilterIndicator(data)
-		container.candidateFilters = candidate
+		local candidate = E:Auras_FilterSlot(container, data)
+		E:Auras_CleanCandidates(container, candidate)
 
 		local slotFilter = container.filter .. (data.anyUnit and '' or '|PLAYER')
 		if container.known[key] then
@@ -734,6 +778,24 @@ function E:Auras_SetIndicator(container)
 	end
 end
 
+function E:Auras_CleanClone(container, key, data)
+	local clone = container.clones[key]
+	if not clone then
+		clone = {}
+		container.clones[key] = clone
+	else
+		wipe(clone)
+	end
+
+	return E:CopyTable(clone, data)
+end
+
+function E:Auras_CleanCandidates(container, data)
+	wipe(container.candidates)
+
+	return E:CopyTable(container.candidates, data)
+end
+
 function E:Auras_SetupList(container, auraTable)
 	wipe(container.keys)
 
@@ -741,15 +803,16 @@ function E:Auras_SetupList(container, auraTable)
 		local key = spell..''
 		if container.isIndicator then
 			if data.enabled then
-				container.keys[key] = data
+				container.keys[key] = E:Auras_CleanClone(container, key, data)
 			end
 		elseif container.isHighlight then
 			if data.enable then
-				if not data.id then
-					data.id = spell
+				local clone = E:Auras_CleanClone(container, key, data)
+				if not clone.id then
+					clone.id = spell
 				end
 
-				container.keys[key] = data
+				container.keys[key] = clone
 			end
 		end
 	end
@@ -887,7 +950,7 @@ function E:Auras_SetContainer(container)
 	E:Auras_SetFlowLayout(container)
 
 	for key, filter in next, container.active do -- known but not active anymore
-		if container.known[key] and (container.filters[key] ~= filter) then
+		if container.filters[key] ~= filter then
 			container:SetAuraGroupMaxFrameCount(key, 0)
 		end
 
@@ -936,23 +999,26 @@ end
 
 function E:Auras_SetUnit(container, unit)
 	container:SetUnit(unit)
-
 	container.unit = unit
-	container.canAssist = UnitCanAssist('player', unit)
 end
 
 function E:Auras_SetEnabled(container)
 	container:SetEnabled(container.enabled and container.canAssist)
 end
 
-function E:Auras_GroupUnit(container, unit)
-	if not container then return end
-
-	E:Auras_SetUnit(container, unit)
+function E:Auras_AssistUnit(container, unit)
+	container.canAssist = UnitCanAssist('player', unit)
 
 	if container.isHighlight then
 		E:Auras_SetEnabled(container)
 	end
+end
+
+function E:Auras_GroupUnit(container, unit)
+	if not container then return end
+
+	E:Auras_SetUnit(container, unit)
+	E:Auras_AssistUnit(container, unit)
 end
 
 function E:Auras_GetFilter(obj, key)
@@ -982,6 +1048,8 @@ function E:Auras_Create(parent, which, override)
 	container.known = {} -- both
 	container.keys = {} -- indicators
 	container.indicators = {}
+	container.clones = {}
+	container.candidates = {} -- only for debug
 	container.active = {} -- groups
 	container.buttons = {}
 	container.layout = {}
@@ -991,11 +1059,11 @@ function E:Auras_Create(parent, which, override)
 	events.owner = container
 	container.events = events
 
-	-- bugged: vehicle
-	events:RegisterEvent('UNIT_FACTION')
-	events:RegisterEvent('UNIT_TARGETABLE_CHANGED')
-
 	-- aurabar to switch to friendship
+	events:RegisterEvent('UNIT_FACTION') -- highlight: faction changes
+	events:RegisterEvent('UNIT_FLAGS') -- highlight: flags changes
+	events:RegisterEvent('UNIT_PHASE') -- highlight: phase changes
+	events:RegisterEvent('GROUP_ROSTER_UPDATE')
 	events:RegisterEvent('PLAYER_TARGET_CHANGED')
 	events:RegisterEvent('PLAYER_FOCUS_CHANGED')
 	events:SetScript('OnEvent', E.Auras_OnEvent)